refactor(sip): extract the fork/group/dial-plan engine into CallForker
Move buildInviteFork, startBroadcastFork, huntRingNext, redirectInvite,
buildCancel, routePageZone, routeRingGroup, and routeDialPlan out of
RequestsHandler into CallForker -- the code that turns a resolved routing
decision into forked, redirected, or cancelled INVITE traffic. Fifteen
call sites across onInvite, onBusy, onUnavailable, onRefer, tick(), and
onPickup become _forker.xxx(...). The 999 target-selection loop stays in
onInvite and passes its vector in, as before.

CallForker takes PbxEnv& (enqueue/localIp/serverPort/findRegistered/
findSession/allocSession/insertSession/contactFor -- all pre-existing
virtuals, none added), PbxFeatureConfig& for the dial-plan/group/zone
lookups, and ParkOrbit& because routeDialPlan's ParkOrbit action calls
_park.orbitIndex()/onInvite() directly -- a sibling reference, not a new
virtual. PbxEnv is unchanged.

routeRingGroup's empty-members UNAVAILABLE now enqueues to the INVITE's
source directly, matching routePageZone's identical branch; endHandle's
404 fallback was unreachable there because caller is already resolved.
NO_ANSWER_TIMEOUT moves to pbx::kNoAnswerTimeout in PbxConfig.hpp -- both
huntRingNext (moving) and onInvite's CFNA arm (staying) use it. Also
updates PbxFeatureConfig.hpp's dialPlan() comment, which pointed at
routeDialPlan as "still in RequestsHandler.cpp until the fork/group
engine extraction."

startPaging, handlePagingAnswer, and buildPagingBye are declared but
never defined or called; they are left in place for a separate cleanup.

Host 309/309 (excl. one pre-existing flaky test), firmware compiles for
SIP_TRANSPORT=eth, callgraph guard clean. Adversarial verify: five lenses
including pool-refusal-contract, all clean; one flagged the
PbxFeatureConfig.hpp comment (fixed before commit).
