+++ tests/CMakeLists.txt
@@ -176,6 +176,9 @@ add_executable(sip_parser_tests
     # The DTMF digit-collection state machine + CLASS/admin codes, extracted
     # out of RequestsHandler into its own translation unit.
     ../src/SIP/DtmfFeatureCodes.cpp
+    # The fork/group/dial-plan routing engine, extracted out of RequestsHandler
+    # into its own translation unit.
+    ../src/SIP/CallForker.cpp
     ../src/SIP/RequestsHandler.cpp
     ../src/SIP/TransactionLayer.cpp
     ../src/SIP/Registrar.cpp
